Seperate extension tests
