+++ tests/web_api/basic_case/assets/test_helpers_with_stop_date.yaml
