+++ tests/web_api/basic_case/test_variables.py
